feature/93 - adds airport/runway dir and move runway files there * updates tests and imports to account for new dir * removed dead code in AirportModel
