+++ src/assets/scripts/client/airport/runway/RunwayCollection.js
@@ -1,11 +1,11 @@
 import _find from 'lodash/find';
 import _forEach from 'lodash/forEach';
 import _isArray from 'lodash/isArray';
-import BaseCollection from '../base/BaseCollection';
+import BaseCollection from '../../base/BaseCollection';
 import RunwayModel from './RunwayModel';
 import RunwayRelationshipModel from './RunwayRelationshipModel';
-import { extrapolate_range_clamp } from '../math/core';
-import { degreesToRadians } from '../utilities/unitConverters';
+import { extrapolate_range_clamp } from '../../math/core';
+import { degreesToRadians } from '../../utilities/unitConverters';
 
 /**
  * Collection of `RunwayModel`s
